QuestionService에 Result를 옮기는 리팩토링 작업 진행

+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -1,12 +1,15 @@
 package com.codessquad.qna.service;
 
 import com.codessquad.qna.domain.Question;
+import com.codessquad.qna.domain.Result;
 import com.codessquad.qna.domain.User;
 import com.codessquad.qna.exception.NoQuestionException;
 import com.codessquad.qna.exception.NoUserException;
 import com.codessquad.qna.repository.QuestionRepository;
+import com.codessquad.qna.web.HttpSessionUtils;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Service
@@ -24,22 +27,47 @@ public class QuestionService {
         questionRepository.save(question);
     }
 
-    public void delete(Question question) {
+    public Result delete(Long id, boolean isLogin, User sessionUser) {
+        Question question = getQuestionById(id);
+        Result result = valid(isLogin, sessionUser, question);
+        if (!result.isValid()) {
+            return result;
+        }
         questionRepository.delete(question);
+        return result;
     }
 
     public List<Question> getQuestionList() {
         return questionRepository.findAll();
     }
 
-    public Question getQuestionById(long id) {
+    public Question getQuestionById(Long id) {
         return questionRepository.findById(id).orElseThrow(NoQuestionException::new);
     }
 
-    public void updateQuestion(long id, Question updateQuestion) {
-        Question question = questionRepository.findById(id).orElseThrow(NoQuestionException::new);
+    public void updateQuestion(Long id, Question updateQuestion) {
+        Question question = getQuestionById(id);
         question.update(updateQuestion);
         questionRepository.save(question);
+    }
+
+    public Result valid(boolean isLoginUser, User sessionUser, Question question) {
+        if (!isLoginUser) {
+            return Result.fail("로그인을 먼저 진행해주세요.");
+        }
+
+        if (!question.isMatchingWriter(sessionUser)) {
+            return Result.fail("수정할 수 있는 권한이 없습니다.");
+        }
+
+        return Result.ok();
+    }
+
+    public Result valid(boolean isLoginUser) {
+        if (!isLoginUser) {
+            return Result.fail("로그인을 먼저 진행해주세요.");
+        }
+        return Result.ok();
     }
 
 }
